Correction of typos in the names of authors, and minor bug fixes
Rectified the names of the qaris in the file LnaguageClass, in
particular adding Capital letters to thei first/last names. In addition
I corrected a minor bug concerning the qaris identified by 'mohna' and 'kafi'
which had the same name displayed on the app. Furthermore, I corrected typos
in the surahs' names (in the same file).
Besides, I added the Qari Omar Al-Kazabri (Warsh), the audio source of which can
be found in the same server.
Finally, I modified the README file to explain what the user is able to do.